edutopia/requirements.txt
Django==4.2.13
groq==0.9.0
opencv_python==4.10.0.84
pdfminer.six==20231228
Pillow==10.4.0
pytesseract==0.3.10
python_pptx==0.6.18
